fixed issue with reading the last word twice
"!my_file.eof(), "is not returning false soon enough." It will still be true after you have read the last item in the file. It only becomes false when you attempt to read one item past the end of the file.

All the various ways of asking the question "have I reached the end of the file?" in C and C++ have the same semantics. This is because the underlying OS primitive, read, cannot return data and signal EOF at the same time."